Extract class SMTFailureTactic
This class was embedded into SMTProversCore with no good reason. It is
now a top-level type.

Also, it has been transformed into an Enum class
as it was already having an Enum-like interface. Finally, remove methods
that return one of the Enum value.

Move the class to the UI plug-in which is the only one to use it.

Use static imports where possible.

git-svn-id: svn://svn.code.sf.net/p/rodin-b-sharp/svn/trunk@14306 1434b563-b632-4741-aa49-43a3a8374d2e