Move Translation preferences to the main SMT preference page
The main SMT preference page was empty, which looks odd.  I've therefore
moved the contents of the Translation preference page to the main
preference page and suppressed the Translation page.

Also added that the translation directory must not be an empty string.
We still allow an empty string for the veriT binary, as it is not
required when using PP translation.