Implement enablement preference in SMT UI plug-in
Move the configuration enablement preference to the SMT UI plug-in, as
this preference is not relevant to the core operation, but only to the
UI plug-in (it gives the list of solvers that shall be run when clicking
the SMT button in the GUI).

The configuration preference page is now fully connected to the SMT UI
preference store.