Generalize field editor for executable file
Make the class used to enter the path to the veriT translator more
general so that it can be reused in other places for an arbitrary
executable file whose path is entered by the user.

Also, trim the text input by the user before entering it in the
preference (as it is trimmed when it is verified).