Move default values in class AutoTactics.
These default values are used only in class AutoTactics in plug-in code.
The tests happen to use the same default value, but there is no real
reason for that.  Therefore, I have duplicated these values in the test
plug-in so that later they can change without impacting the core
plug-in (decoupling).