Fix performance tests.
Ensure that tests succeed or are ignored if the binary file of the
SMT solver is not available.  The exact display in Eclipse will depend
on the resolution of Eclipse bug #359944.

Ignore tests in classes that extend an acceptance test (these tests can
not run an unbundled solver).

Make class UnsatCoreExtractionPerfWithPP abstract as it is not intended
to be run standalone.

Now, all tests are green when no external solver is available.