Break dependence of performance tests on UI plug-ins.
To run tests on real projects, there was a dependence onto the SMT UI
plug-in.  However, the code for setting the configurations is too old
and is not working anymore.  I have therefore commented out and signaled
the problem with a FIXME.

This will be repaired when there will be need again to run performance
tests on real projects.

Also remove unneeded dependence on AST plug-in.