Move the tactic parameterizer back in the core
In hindsight, it was a bad idea to move the tactic parameterizer into
the UI plug-in. It really belongs to the core plug-in, as it has no
dependency on any graphical part.

Change the API of the core plug-in to only expose tactic descriptors
built from the tactic parameterizer.

Update code of the UI plug-in to this new API.

Add simple tests for the tactic parameterizer class, including a test
that shows that tactic descriptors returned by the core plug-in are
indeed usable to discharge a sequent.

There are still some loose ends about when the existence of an SMT
configuration should be checked.  This will be fixed in a later commit.