Merge bundled and user-defined descriptors
Provide one unified method for accessing solver and configuration
descriptors.  The API does not distinguish anymore between bundled
and user-defined descriptors but returns all together.  This permits
to maintain an order amongst descriptors of the same kind.

Update the UI plug-in to this new API.

Add tests to ensure that bundled descriptors cannot be removed.