Move configuration enablement to core
The fact that some configuration is enabled was persisted in the UI
plug-in.  However, we would like to move the construction of the default
SMT auto-tactic to the core plug-in. We therefore move the code for
managing enablement of SMT configurations to the core plug-in.  The
enabled attribute is now persisted together with the rest of the
configuration descriptor, rather than in a separate list of names.

Another advantage of this new solution is that the order of the
configurations and their enablement is now stored in the same place,
including the place of disabled configurations, which was not the case
with the previous solution.