Move SMT auto-tactic to core
The core plug-in now contributes a tactic for running all enabled SMT
configurations on the current sequent. This is a much better place than
in the UI, now that all preferences are stored in the core plug-in.

Update UI plug-in to use the core tactic.

Add acceptance test for the new auto tactic.