The SMT reasoner takes a configuration name as input
The SMT reasoner used to take a full SMT configuration descriptor as
input, but this caused early binding.  Moreover, it did not correspond
to what was actually serialized, that is the configuration name alone.

Instances of SMTInput are now built with the configuration name and
resolve this name to a known configuration in a lazy manner.

This solution is not perfect (I would have preferred that the reasoner
itself performs the resolution) but it is the best we can do with the
XProver API of Rodin 2.7.

Update code that creates instances of SMTInput.

Add corresponding tests for class SMTInput.