Cleanup UI plug-in
Make the UI use the auto-tactic from the core plug-in.

Remove the tactic profile for running all enabled SMT solvers, which
is replaced by the SMT auto-tactic.

A visible change is that the SMT auto-tactic is no longer run within
an attempt after lasso from the UI.  This was not a good idea as it was
not truthful to the end-user.  If needed, end-users can always define
a tactic profile to add the lasso.

Remove all cruft that accumulated and is no longer needed.

Rename class SMT to SMTTacticProvider which is more meaningful.

Extract code for installing the default auto + SMT profile into the new
class TacticProfileContribution.