Move tactic building code in core plug-in
Move the code computing the "Default Auto Tactic with SMT" into the SMT
core plug-in instead of the UI plug-in.

This introduces a new dependency from the SMT core plug-in to the
Event-B core plug-in (to have direct access to the default Auto tactic).

The code has been streamlined and made more robust during the move.

Add an acceptance test for this tactic.