Automate specification of SMT-LIB version
Most solvers take some parameters that specify the SMT-LIB version of
their input benchmarks.  Until now, these parameters had to be specified
by the user in the SMT configuration, which was cumbersome and
error-prone when changing the SMT language version.

This is now automated by the SMT solver plug-in which contains
hard-coded parameters to pass to the solver depending on the language
version of the generated benchmark.