Simplify intermediate file management
Extract code manipulating temporary files into separate class.

Do not create temporary directories that are not needed.

Do not create temporary files that are not needed (i.e. result file).

Delete files as soon as they are not needed anymore rather than when
closing the platform.  This is because the platform can run for a long
time and create a lot of files which would clutter the user's disk.

Remove instance locking which is useless.

Do not use fancy names for temporary files: it is more dangerous than
useful (in case the lemma name happens to have some special meaning for
the host operating system).