Update build scripts to use Git
The build scripts were designed for a Subversion repository.  This
commit allows to fetch sources from a Git repository instead.

The version qualifier used to be a subversion revision number, it is now
a short Git commit identifier (short SHA-1).

Remove leading "r" in version qualifier as this is Subversion specific
and not needed anymore.

Make sure that the release engineer sets the appropriate Git commit.