Use a Rodin installation as target platform
Originally the Eclipse used to build the plug-ins and the target
platform were the same.  This is no longer the case (because the Rodin
editor adds a dependency onto EMF plug-ins which are not installed in
the Eclipse used to build plug-ins). So we put the correct path to the
target platform in the build script.