Add tracing option for tactics
This new option allows to trace the updates and launches of the SMT
Auto tactic.  As a consequence, the initialization of the list of
tactics to run is now delayed from class creation to its first use.
This was needed to avoid an attempt to initialize before the plug-in
was fully started (e.g., because the start() method of the plug-in
would load the SMTAutoTactic class to set its DEBUG flag).