Make CVC4 and Z3 features.
The aim is to make them optional within the main SMT Feature.
It's the best solution to deal with the absence of support of Linux 32 bit OSes.

Indeed, we cannot put a platform filter on the global SMT feature.
Thus, we put a platform filter on the z3 and cvc4 plug-ins.
But then, if we do just that, as the included plug-ins are always
strictly required, when trying to install on Linux 32 bits,
it's impossible and the reason is a non obvious conflict.
So we include CVC4 and Z3 as optional features, making it possible to install
the global feature without them, so as to update VeriT and CVC3, which are
released as 32 bit binaries.