veriT: make a feature
The main motivation is to be able to correctly filter
the supported architectures at the feature level.
Thus, the filtering occurs at install time and the user
sees exactly which prover extensions are going to be
installed.

It is also a good way to structure the plug-in hierarchy
and to provide specific licenses.
Features already existed for Z3 and CVC4.