Ensure proper proof rule dependencies
If the solver fails to produce an unsat core, although it is expected
to produce one, the SMT integration used to produce a proof rule with
no dependency at all, which is meaningless and an error.

We now protect against this case by checking whether the unsat core is
empty.