Bundles extension point: set 'library' kind to 'resource'
This indication makes the extension editor add a "Browse" button
to ease the selection of the library within the plug-in.
The plug-in validator also detects a problem if the path is invalid.
The validator does not understand the "$os$" variable though,
so the bare path must be given; however, using the "Browse" button
makes a valid path.